+++ pkg/analysis_server/lib/src/lsp/handlers/code_actions/pubspec.dart
@@ -7,9 +7,9 @@ import 'dart:async';
 import 'package:analysis_server/lsp_protocol/protocol.dart';
 import 'package:analysis_server/src/lsp/handlers/code_actions/abstract_code_actions_producer.dart';
 import 'package:analysis_server/src/services/correction/fix/pubspec/fix_generator.dart';
+import 'package:analyzer/source/file_source.dart';
 import 'package:analyzer/source/line_info.dart';
 import 'package:analyzer/src/pubspec/pubspec_validator.dart';
-import 'package:analyzer/src/source/source_resource.dart';
 import 'package:yaml/yaml.dart';
 
 /// Produces [CodeAction]s from Pubspec fixes.